update Events class to pull data from django backend
